Revert "[nrf fromtree] net: mgmt: Convert the mgmt API to use 64-bit masks"
This reverts commit 8b3dda735e1b57360d80423698f4897c7a790c8b.

+++ drivers/modem/modem_cellular.c
@@ -1774,7 +1774,7 @@ static int modem_cellular_pm_action(const struct device *dev, enum pm_device_act
 }
 #endif /* CONFIG_PM_DEVICE */
 
-static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
+static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
 				   struct net_if *iface)
 {
 	struct modem_cellular_data *data =
