+++ drivers/wifi/esp32/src/esp_wifi_drv.c
@@ -77,7 +77,7 @@ struct esp32_wifi_runtime {
 
 static struct net_mgmt_event_callback esp32_dhcp_cb;
 
-static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
+static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
 			       struct net_if *iface)
 {
 	switch (mgmt_event) {
